refactor: settings
