+++ next-docs/components/Layout.tsx
@@ -1,39 +1,9 @@
 import React, { useState } from 'react';
 import Breadcrumbs from './breadcrumbs/Breadcrumbs';
 import Footer from './Footer';
+import ReactSettings from './settings/react/ReactSettings';
 import Sidebar from './sidebar/Sidebar';
 import SidebarTransition from './sidebar/SidebarTransition';
-import BrandThemeSelector from './themes/BrandThemeSelector';
-import DarkLightModeSwitcher from './themes/DarkLightModeSwitch';
-import { useDocsTheme } from './themes/DocsThemeProvider';
-import RTLModeSwitch from './themes/RTLModeSwitch';
-
-const brandMap = {
-  moonDesign: 'moon',
-  sportsbet: 'sportsbet',
-  bitcasino: 'bitcasino',
-  betadda: 'betadda',
-  // hub88: 'hub88', <-- previously was used for the old design
-};
-
-export type Brand = keyof typeof brandMap | null;
-interface GetClassNameProps {
-  brand: Brand | null;
-  colorMode: string | null;
-}
-
-const singleThemedBrands: string[] = [];
-
-const getClassName = ({ brand, colorMode }: GetClassNameProps) => {
-  const brandName = brand ? brandMap[brand] : '';
-  const singleThemed = singleThemedBrands.includes(brandName);
-  if (brandName) {
-    return singleThemed
-      ? `theme-${brandName}`
-      : `theme-${brandName}-${colorMode}`;
-  }
-  return 'theme-moon-dark';
-};
 
 interface LayoutProps {
   children: React.ReactNode;
@@ -41,31 +11,11 @@ interface LayoutProps {
 
 export default function Layout({ children }: LayoutProps) {
   const [sidebarOpen, setSidebarOpen] = useState(false);
-  const { toggleColorScheme, getColorMode, setBrand, getBrand, themeKeys } =
-    useDocsTheme();
   const openSidebar = () => setSidebarOpen(true);
-  const className = getClassName({
-    brand: getBrand() as Brand,
-    colorMode: getColorMode(),
-  });
-
-  const [isRtl, toggleDir] = useState(false);
-
-  const toggleDirection = () => {
-    const htmlTag = document && document?.getElementsByTagName('html')[0];
-    toggleDir(!isRtl);
-    const dirAttr = htmlTag.getAttribute('dir');
-
-    if (!dirAttr) {
-      htmlTag.setAttribute('dir', 'rtl');
-    } else {
-      htmlTag.removeAttribute('dir');
-    }
-  };
 
   return (
     <div
-      className={`${className} support-colors bg-background flex rounded-tl-3xl rounded-tr-3xl`}
+      className={`support-colors bg-background flex rounded-tl-3xl rounded-tr-3xl`}
     >
       {/* Dynamic sidebar with transition for mobile */}
       <SidebarTransition isOpen={sidebarOpen} setIsOpen={setSidebarOpen}>
@@ -111,23 +61,8 @@ export default function Layout({ children }: LayoutProps) {
           <Breadcrumbs />
         </div>
 
-        {/* TODO overflow-y-auto */}
         <main className="flex flex-col flex-1 relative overflow-y-auto focus:outline-none">
-          <BrandThemeSelector
-            themeKeys={themeKeys}
-            setBrand={setBrand}
-            isRtlEnabled={isRtl}
-            darkLight={
-              <DarkLightModeSwitcher
-                toggle={toggleColorScheme}
-                isEnabled={getColorMode() === 'dark'}
-                isRtlEnabled={isRtl}
-              />
-            }
-            rtlSwitch={
-              <RTLModeSwitch toggle={toggleDirection} isEnabled={isRtl} />
-            }
-          />
+          <ReactSettings />
 
           <div className="py-6 px-4 md:px-16 lg:px-0">{children}</div>
