+++ next-docs/components/settings/react/DarkLightModeSwitch.tsx
