+++ next-docs/components/settings/react/RTLModeSwitch.tsx
